Rename assignment-choose and project-choose
