+++ server/front/src/app/views/shared/assignment-choose/assignment-choose.component.html
