+++ server/front/src/app/views/shared/commit-record/commit-record.component.html
@@ -63,7 +63,7 @@
     <span><a>
         <h3>Feedback Information {{commitNumber}}</h3>
       </a></span>
-    <table class='table table-responsive-sm table-bordered table-hover table-outline mb-0 text-center'>
+    <table class='table table-responsive-sm table-bordered table-hover table-outline mb-0'>
       <thead class="thead-light">
         <tr>
           <th class='w-10'>Line</th>
