+++ server/front/src/app/views/shared/project-choose/project-choose.component.html
