+++ server/front/src/app/views/shared/shared.module.ts
@@ -8,8 +8,8 @@ import { FormsModule } from '@angular/forms';
 import { GroupFilterPipe } from './../../pipe/group-filter.pipe';
 import { FilterPipe } from './../../pipe/filter.pipe';
 
-import { AssignmentChoosedComponent } from '../shared/assignment-choosed/assignment-choosed.component';
-import { ProjectChoosedComponent } from './project-choosed/project-choosed.component';
+import { AssignmentChoosedComponent } from './assignment-choose/assignment-choose.component';
+import { ProjectChoosedComponent } from './project-choose/project-choose.component';
 import { GroupDashboardComponent } from '../teacher/group-dashboard/group-dashboard.component';
 import { CommitRecordComponent } from './commit-record/commit-record.component';
 import { DashboardStatusComponent } from './dashboard-status/dashboard-status.component';
